+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* feat: --zones scope and use CCM node metadata by default #143
+
 ## v0.33.1
 * feat(node): add preStop hook #132